+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 [Full changelog](https://github.com/mozilla/glean/compare/v44.1.1...main)
 
+* General
+  * The `glean.error.preinit_tasks_overflow` metric now reports only the number of overflowing tasks.
+    It is marked as version 1 in the definition now. ([#2026](https://github.com/mozilla/glean/pull/2026))
 * Kotlin
   * (Development only) Allow to override the used `glean_parser` in the Glean Gradle Plugin ([#2029](https://github.com/mozilla/glean/pull/2029))